+++ src/commands/moveStalledJobsToWait-9.lua
@@ -28,6 +28,7 @@ local rcall = redis.call
 --- @include "includes/batches"
 --- @include "includes/decreaseConcurrency"
 --- @include "includes/getTargetQueueList"
+--- @include "includes/moveParentFromWaitingChildrenToFailed"
 --- @include "includes/removeJob"
 --- @include "includes/removeJobsByMaxAge"
 --- @include "includes/removeJobsByMaxCount"
@@ -83,7 +84,9 @@ if (#stalling > 0) then
                     local stalledCount =
                         rcall("HINCRBY", jobKey, "stalledCounter", 1)
                     if (stalledCount > MAX_STALLED_JOB_COUNT) then
-                        local rawOpts = rcall("HGET", jobKey, "opts")
+                        local jobAttributes = rcall("HMGET", jobKey, "opts", "parent")
+                        local rawOpts = jobAttributes[1]
+                        local rawParentData = jobAttributes[2]
                         local opts = cjson.decode(rawOpts)
                         local removeOnFailType = type(opts["removeOnFail"])
                         rcall("ZADD", failedKey, timestamp, jobId)
@@ -95,6 +98,16 @@ if (#stalling > 0) then
                               "failed", "jobId", jobId, 'prev', 'active',
                               'failedReason', failedReason)
 
+                        if opts['fpof'] and rawParentData ~= false then
+                            local parentData = cjson.decode(rawParentData)
+                            moveParentFromWaitingChildrenToFailed(
+                                parentData['queueKey'],
+                                parentData['queueKey'] .. ':' .. parentData['id'],
+                                parentData['id'],
+                                jobKey,
+                                timestamp
+                            )
+                        end
                         if removeOnFailType == "number" then
                             removeJobsByMaxCount(opts["removeOnFail"],
                                                   failedKey, queueKeyPrefix)